src/modules/requirements.txt
gdown==4.7.1
pandarallel==1.6.5
pandas==1.5.3
python_Levenshtein==0.21.1
sentence_transformers==2.2.2
streamlit==1.26.0
torch==2.1.0
urllib3==1.26.16
